Move fermi components around
