+++ fulqrum/core/includes/fermi_header.pxi
@@ -12,8 +12,12 @@
 
 include "base_header.pxi"
 
-cdef extern from "../src/fermi.hpp":
-    void jw_term(const FermionicTerm_t& fermi_term, OperatorTerm_t& qubit_term)
+cdef extern from "../src/fermi_term.hpp":
+    
+    void jw_term(const FermionicTerm_t& fermi_term, OperatorTerm_t& qubit_term) nogil
+
+
+cdef extern from "../src/fermi_oper.hpp":
 
     void extended_jw_transform(const FermionicOperator_t& fermi, QubitOperator_t& out,
-                               size_t num_terms)
+                               size_t num_terms) nogil
