+++ fulqrum/core/src/fermi_term.hpp
@@ -13,6 +13,7 @@
  */
 #pragma once
 #include "constants.hpp"
+#include "qubit_term.hpp"
 #include <algorithm>
 #include <cmath>
 #include <complex>
@@ -114,3 +115,87 @@ typedef struct FermionicTerm
         coeff *= prefactor;
     }
 } FermionicTerm_t;
+
+
+/**
+ * Compute the JW phase for a given operator
+ *
+ * @param[in] op The operator in question
+ * 
+ * @return Integer phase value
+ */
+inline int jw_phase(const unsigned char op)
+{
+    int out;
+    switch (op)
+    {
+    case 5: //minus sign if op = '-'
+        out = -1;
+        break;
+    case 2: //minus sign if op = '1'
+        out = -1;
+        break;
+    default:
+        out = 1;
+        break;
+    }
+    return out;
+}
+
+
+/**
+ * Compute the extended JW transformation for a single Fermionic term
+ *
+ * @param[in] fermi_term Input Fermionic term
+ * @param[in,out] qubit_term Output qubit term
+ */
+void jw_term(const FermionicTerm_t& fermi_term, OperatorTerm_t& qubit_term)
+{
+    int num_elems = fermi_term.indices.size();
+    int kk, mm;
+    unsigned int jj;
+    int phase = 1;
+    unsigned int current_ind;
+    unsigned char current_val;
+    qubit_term.coeff = fermi_term.coeff;
+    qubit_term.extended = (num_elems > 0);
+    //Start with do_z = 0 since nothing has been done yet
+    int do_z = 0;
+    for(kk = num_elems - 1; kk > -1; kk--)
+    {
+        current_ind = fermi_term.indices[kk];
+        current_val = fermi_term.values[kk];
+        // Add start element to qubit operator
+        qubit_term.indices.push_back(current_ind);
+        qubit_term.values.push_back(current_val);
+        // If a Z term acts on the current value then need to account
+        // for the phase factor in the coefficient
+        if(do_z)
+        {
+            phase *= jw_phase(current_val);
+        }
+        // update do_z with this operator
+        do_z ^= (current_val > 4);
+        // if not at last element in num_elems and do_z
+        // make every id element between start and the next elem a Z operator
+        if(kk && do_z)
+        {
+            for(jj = current_ind - 1; jj > fermi_term.indices[kk - 1]; jj--)
+            {
+                qubit_term.indices.push_back(jj);
+                qubit_term.values.push_back(0);
+            }
+        }
+        // If only one element exists then kk=0 but I still need to
+        // add Z operators down to zero
+        else if(num_elems == 1 && do_z)
+        {
+            for(mm = current_ind - 1; mm > -1; mm--)
+            {
+                qubit_term.indices.push_back(mm);
+                qubit_term.values.push_back(0);
+            }
+        }
+    } // end kk loop
+    qubit_term.coeff *= phase; // multiple coefficient by phase factor
+}
